move `Swarm.Game.Log` -> `Swarm.Log`
It's no longer specific to "game" since we use it for system logs as
well as robot logs.
