+++ test/unit/TestNotification.hs
@@ -10,9 +10,9 @@ import Control.Lens (Getter, Ixed (ix), view, (&), (.~), (^.), (^?!))
 import Data.Text (Text)
 import Data.Text qualified as T
 import Swarm.Game.CESK (TickNumber (..))
-import Swarm.Game.Log
 import Swarm.Game.Robot
 import Swarm.Game.State
+import Swarm.Log
 import Test.Tasty
 import Test.Tasty.HUnit
 import TestUtil
